Move mechanism module (non-breaking)
This moves the mechanism module out of types and into
the location indicated in commit 20b26a5.

src/types/mechanism/* -> src/mechanism/*
src/types/mod.rs::MechanismFlags -> src/mechanism/flags.rs
