+++ cryptoki/src/mechanism/rsa.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //! RSA mechanism types
 
-use crate::types::mechanism::{Mechanism, MechanismType};
+use super::{Mechanism, MechanismType};
 use crate::types::Ulong;
 use crate::{Error, Result};
 use cryptoki_sys::*;
